Modify card height styles and footer hight
Change height form fit-content, to .content 1fr, for the grid to adjust it's size to the biggest element.
To limit the first row height I added a 1fr row. 1fr on auto rows for consistency.

Footer hight was changed from a relative size to a fixed auto size, while adding a padding on the element itself.
This change is introduced since the footer got bigger with the addition of tasks.