Update Styles; Add task-complete logic
New styles for `task completed`, and `task title hover`.
Change footer color to `color-primary` and font color to `color-primary-light`

task-complete logic was possible by associating a task to an id.
A `Map` was probably a better choice to use, but for this one I've decided to keep the array since converting the code to fit the `Map` syntax showed to be more time consuming and introduced new bugs.