Modify util/Render pages/activeTasks, Add pages/completeTasks
Change render from accepting HTML nodes to accepting callback functions that return nodes.
This is done to be able to refresh the main pages upon some event or change.

The activeTasks page is modified to fit the modified render function.

completeTasks page is updated to only show done tasks in realtime.